자신이 작성한 리뷰 상세 조회

+++ src/main/java/com/example/block/dto/ReviewResponseDTO.java
@@ -20,7 +20,7 @@ public class ReviewResponseDTO {
         String content;
         String service;
         String prize;
-        Float score;
+        Double score;
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
         LocalDateTime createdAt;
     }
